AtomicInt now used for _cputime

+++ src/moa/classifiers/Multithreading.java
@@ -3,12 +3,13 @@ package moa.classifiers;
 import java.util.HashSet;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ForkJoinPool;
+import java.util.concurrent.atomic.AtomicInteger;
 
 public interface Multithreading {
 
      void trainingHasEnded();
 
-     double getCpuTime();
+     AtomicInteger getCpuTime();
 
      void init() throws InterruptedException, ExecutionException;
